+++ packages/portal/spec-renderer/src/components/SpecDetails.vue
@@ -1,10 +1,14 @@
 <template>
-  <div class="kong-portal-spec-details">
+  <div
+    class="kong-public-ui-spec-details"
+    :class="{ 'active-op-focused': activeOperationFocused }"
+  >
     <kong-swagger-ui
       v-if="hasRequiredProps"
       ref="swaggerRef"
       :application-registration-enabled="applicationRegistrationEnabled"
       :current-version="currentVersion"
+      data-testid="kong-public-ui-spec-details-swagger"
       :essentials-only="essentialsOnly"
       :has-sidebar="hasSidebar"
       :relative-sidebar="relativeSidebar"
@@ -13,7 +17,7 @@
     />
     <div
       v-else
-      data-testid="kong-portal-spec-details-error"
+      data-testid="kong-public-ui-spec-details-error"
     >
       {{ i18n.t('specDetails.error') }}
     </div>
@@ -38,7 +42,7 @@ const props = defineProps({
   },
   hasSidebar: {
     type: Boolean,
-    default: true,
+    default: false,
   },
   relativeSidebar: {
     type: Boolean,
@@ -70,13 +74,16 @@ const hasRequiredProps = computed((): boolean => {
   return !!(props.document || props.url)
 })
 
-function showAndScrollToOperation() {
+const activeOperationFocused = ref(false)
+const showAndScrollToOperation = (): void => {
   if (!swaggerRef.value) {
     return
   }
 
   swaggerRef.value.showOperation(props.activeOperation)
   swaggerRef.value.scrollToOperation(props.activeOperation)
+
+  activeOperationFocused.value = true
 }
 
 onMounted(() => {
@@ -93,7 +100,7 @@ watch(() => props.activeOperation, () => {
 </script>
 
 <style lang="scss" scoped>
-.kong-portal-spec-details {
+.kong-public-ui-spec-details {
   color: var(--kong-portal-spec-details-text-color, #000);
   font-family: var(--kong-portal-spec-details-font-family-default, Roboto, Helvetica, sans-serif);
   font-size: var(--kong-portal-spec-details-font-size, 16px);